Evaluation class

Run the evaluation workflow

Methods

Public methods

Evaluation$new()

Evaluation$run()

Evaluation$clone()

Evaluation$new()

Create an evaluation workflow

Usage

Evaluation$new(
  config,
  workspace = NULL,
  backend = NULL,
  logger = default_logger,
  summary_class = ComparisonSummary
)

Arguments

config

the configuration of the evaluation workflow

workspace

an object of class EvalWorkspace to access the evaluation data (default: EvalWorkspace$new(config$eval_workspace))

backend

an object of class ParallelBackend to run parallel computations (default: ParallelBackend$new(config$parallel, config$cores))

logger

a logger object with info, debug, warn and error methods (default: uses the logger package)

summary_class

a class to build the summary of the evaluation (default: ComparisonSummary)

Evaluation$run()

Run the evaluation workflow This function orchestrates the full evaluation workflow based on a given configuration object. It initializes logging, optionally prepares the evaluation workspace, runs the evaluation for all species, and displays summary information.

Usage

Evaluation$run()

Evaluation$clone()

The objects of this class are cloneable with this method.

Usage

Evaluation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
config <- Configuration$new(
  stics_exe = "/path/to/stics",
  eval_workspace = "workspace/",
  metadata_file = "metadata.csv"
)
Evaluation$new(config)$run()
} # }