Running simulations of STICS using a STICS executable

Running simulations of STICS using a STICS executable

run_simulations(
  stics_exe,
  workspace,
  usm_names,
  successive,
  verbose,
  parallel = FALSE,
  cores = NA
)

Arguments

stics_exe

path to the STICS executable

workspace

path to the simulation and observation data

usm_names

a list of USM name

successive

a list of vectors containing the names of the UMSs to consider as successive (e.g. list(c("usm1.1","usm1.2"),c("usm2.1","usm2.2")) defines 2 successions usm1.1->usm1.2 and usm2.1->usm2.2)

verbose

a boolean used to run simulation in verbose mode

parallel

Boolean. Is the computation to be done in parallel ?

cores

Number of cores to use for parallel computation.

Value

a list of crop simulation

Examples

  run_simulations(
    "/path/to/stics/exe",
    "/path/to/workspace",
    c("usm1", "usm2", "usm3"),
    TRUE
  )
#> Error in run_simulations("/path/to/stics/exe", "/path/to/workspace", c("usm1",     "usm2", "usm3"), TRUE): could not find function "run_simulations"